+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Update Opsgenie instructions
 - Update Spike.sh instructions
+- Add system check to validate settings.SITE_ROOT (#895)
 
 ## v3.1 - 2023-12-13